requirements.txt
Flask==0.12.2
Flask-PyMongo==0.4.1
Flask-RESTful==0.3.5
gunicorn==19.7.1